[#399] Enable show label in ChartCalculatedHhIncome
