+++ frontend/src/App.scss
@@ -409,6 +409,7 @@
     }
 
     .button-export {
+      margin-left: 5px;
       border: 1px solid #01625f;
       font-weight: 600;
       font-size: 12px;
